await_order_processed 11

+++ internal/model/model.go
@@ -10,18 +10,18 @@ type RegisterSchema struct {
 type OrderSchema struct {
 	Number     string    `json:"number"`
 	Status     string    `json:"status"`
-	Accrual    int       `json:"accrual"`
+	Accrual    float32   `json:"accrual"`
 	UploadedAt string    `json:"uploaded_at"`
 	Date       time.Time `json:"-"`
 }
 
 type AccrualSchema struct {
-	Order   string `json:"order"`
-	Status  string `json:"status"`
-	Accrual int    `json:"accrual"`
+	Order   string  `json:"order"`
+	Status  string  `json:"status"`
+	Accrual float32 `json:"accrual"`
 }
 
 type BalaneSchema struct {
-	Current   int `json:"current"`
-	WithDrawn int `json:"withdrawn"`
+	Current   float32 `json:"current"`
+	WithDrawn float32 `json:"withdrawn"`
 }
